Slightly better error message when API Key isn't right
This shows the full message from the OpenAI API when it is a connection or protocol error. These errors aren't super-clear, but they are better than the HTTP codes that were coming up before.
To try this out, try talking to Molly with no API Key entered, and then enter a valid key that has no money on it (I had to set up a separate, unfunded openai account to test this.)